app/requirements.txt
pandas
panel
azure-storage-blob
tables
jinja2
